feat: connection to survey
