+++ app/auth/callback/route.ts
@@ -4,9 +4,12 @@ import { type CookieOptions, createServerClient } from '@supabase/ssr'
 
 export async function GET(request: Request) {
   const { searchParams, origin } = new URL(request.url)
+  console.log('searchParams', searchParams)
+  console.log('origin', origin)
   const code = searchParams.get('code')
   // if "next" is in param, use it as the redirect URL
   const next = searchParams.get('next') ?? '/'
+  console.log('next', next)
 
   console.log('code', code)
   if (code) {
