+++ app/page.tsx
@@ -42,6 +42,11 @@ export default function Home() {
     <div className="min-h-full mb-24 w-full overflow-visible">
       <Toaster />
       <div className="flex flex-col h-svh relative overflow-visible justify-center ">
+          <div className="hidden sm:mb-8 sm:flex sm:justify-center">
+            <Link href="survey" className="relative rounded-full px-3 py-1 text-sm leading-6 text-gray-600 ring-1 ring-gray-900/10 hover:ring-gray-900/20">
+              We are leading a <span className="text-orange underline">survey</span> to better understand your needs.{' '}
+            </Link>
+          </div>
         <h1 className={cn("text-5xl font-medium text-center", whyte.className)}>
           IS YOUR APP
           <strong className="text-orange"> SECURE?</strong>
@@ -194,21 +199,21 @@ export default function Home() {
           </Accordion>
         </div>
         <div className="bg-white py-24 sm:py-32">
-      <div className="mx-auto max-w-7xl px-6 lg:px-8">
-        <h2 className="text-center text-lg font-semibold leading-8 text-gray-900">
-          
-        </h2>
-        <div className="mx-auto mt-10 grid max-w-lg grid-cols-4 items-center gap-x-8 gap-y-10 sm:max-w-xl sm:grid-cols-6 sm:gap-x-10 lg:mx-0 lg:max-w-none lg:grid-cols-5">
-          <img
-            className="col-span-2 max-h-12 w-full object-contain lg:col-span-1"
-            src="/euratechnologies.png"
-            alt="Euratechnologies"
-            width={158}
-            height={48}
-          />
+          <div className="mx-auto max-w-7xl px-6 lg:px-8">
+            <h2 className="text-center text-lg font-semibold leading-8 text-gray-900">
+
+            </h2>
+            <div className="mx-auto mt-10 grid max-w-lg grid-cols-4 items-center gap-x-8 gap-y-10 sm:max-w-xl sm:grid-cols-6 sm:gap-x-10 lg:mx-0 lg:max-w-none lg:grid-cols-5">
+              <img
+                className="col-span-2 max-h-12 w-full object-contain lg:col-span-1"
+                src="/euratechnologies.png"
+                alt="Euratechnologies"
+                width={158}
+                height={48}
+              />
+            </div>
+          </div>
         </div>
-      </div>
-    </div>
         <div className="flex flex-col justify-center items-center">
           <h2 className={cn("font-medium text-3xl", whyte.className)}>Try our platform now!</h2>
           <p className="text-center text-muted-foreground">
